migrated to cmake
