+++ cmake/InstallFiles/Config.cmake.in
@@ -1,0 +1,15 @@
+@PACKAGE_INIT@
+
+include("${CMAKE_CURRENT_LIST_DIR}/@CMAKE_PROJECT_NAME@.cmake")
+
+set(@CMAKE_PROJECT_NAME@_INCLUDE_DIRS @PROJECT_REQUIRED_INCLUDE_DIRS@)
+set(@CMAKE_PROJECT_NAME_CAP@_INCLUDE_DIRS @PROJECT_REQUIRED_INCLUDE_DIRS@)
+
+set(@CMAKE_PROJECT_NAME@_LIBRARIES @PROJECT_REQUIRED_LIBRARIES@)
+set(@CMAKE_PROJECT_NAME_CAP@_LIBRARIES @PROJECT_REQUIRED_LIBRARIES@)
+
+set(@CMAKE_PROJECT_NAME@_LIBRARY_DIRS @PROJECT_REQUIRED_LIBRARY_DIRS@)
+set(@CMAKE_PROJECT_NAME_CAP@_LIBRARY_DIRS @PROJECT_REQUIRED_LIBRARY_DIRS@)
+
+set(@CMAKE_PROJECT_NAME@_CXX_FLAGS "@PROJECT_REQUIRED_CXX_FLAGS@")
+set(@CMAKE_PROJECT_NAME_CAP@_CXX_FLAGS "@PROJECT_REQUIRED_CXX_FLAGS@")
