+++ cmake/Modules/RTXI.cmake
@@ -1,0 +1,17 @@
+find_package(RTXI REQUIRED)
+
+set(PROJECT_REQUIRED_CXX_FLAGS "${PROJECT_REQUIRED_CXX_FLAGS} ${RTXI_CXX_FLAGS}")
+list(APPEND PROJECT_REQUIRED_INCLUDE_DIRS ${RTXI_INCLUDE_DIRS})
+list(APPEND PROJECT_REQUIRED_LIBRARY_DIRS ${RTXI_LIBRARY_DIRS})
+list(APPEND PROJECT_REQUIRED_LIBRARIES_SHORT_NAME ${RTXI_LIBRARIES})
+list(APPEND CMAKE_SHARED_LINKER_FLAGS ${RTXI_SHARED_LINKER_FLAGS})
+
+# RTXI uses Qt MetaObjectCode (MOC)
+set(CMAKE_AUTOMOC ON)
+
+message(STATUS "RTXI_CXX_FLAGS = ${RTXI_CXX_FLAGS}")
+message(STATUS "RTXI_INCLUDE_DIRS = ${RTXI_INCLUDE_DIRS}")
+message(STATUS "RTXI_LIBRARY_DIRS = ${RTXI_LIBRARY_DIRS}")
+message(STATUS "RTXI_LIBRARIES = ${RTXI_LIBRARIES}")
+message(STATUS "RTXI_SHARED_LINKER_FLAGS = ${RTXI_SHARED_LINKER_FLAGS}")
+
