+++ cmake/Modules/hmm.cmake
@@ -1,0 +1,12 @@
+find_package(hmm REQUIRED)
+
+set(PROJECT_REQUIRED_CXX_FLAGS "${PROJECT_REQUIRED_CXX_FLAGS} ${HMM_CXX_FLAGS}")
+list(APPEND PROJECT_REQUIRED_INCLUDE_DIRS ${HMM_INCLUDE_DIRS})
+list(APPEND PROJECT_REQUIRED_LIBRARY_DIRS ${HMM_LIBRARY_DIRS})
+list(APPEND PROJECT_REQUIRED_LIBRARIES_SHORT_NAME ${HMM_LIBRARIES})
+
+message(STATUS "HMM_CXX_FLAGS = ${HMM_CXX_FLAGS}")
+message(STATUS "HMM_INCLUDE_DIRS = ${HMM_INCLUDE_DIRS}")
+message(STATUS "HMM_LIBRARY_DIRS = ${HMM_LIBRARY_DIRS}")
+message(STATUS "HMM_LIBRARIES = ${HMM_LIBRARIES}")
+
