+++ src/hmm_decoder.cpp.orig
