+++ src/sourcelist.txt
@@ -1,0 +1,1 @@
+hmm_decoder.cpp;
